完成基本的增删改查

+++ 60_Mybatis/src/main/java/com/ljf/bean/Emp.java
@@ -11,7 +11,7 @@ import lombok.Data;
 public class Emp {
     private int id;
     private String name;
-    private byte age;
+    private int age;
     private int deptId;
     private int empNo;
 }
